Ensuring that only Raster Classes (rather than Vectors) can be aggregated by Image instances; Restructuring the namespaces for the physical data models (i. e. moving them out of the GeoConcerns namespace); Extending the RSpec tests for the Solr indexing for RasterFile and VectorFile instances (ensuring that the "crs" field expectations are separate tests); Disintegrating the engine_cart dependencies for testing; Extending the structure for the Gemspec file
