+++ geo_concerns/Rakefile
@@ -8,7 +8,6 @@ require 'bundler/gem_tasks'
 require 'bundler/setup'
 require 'rspec/core/rake_task'
 require 'jettywrapper'
-require 'engine_cart/rake_task'
 require 'rubocop/rake_task'
 
 Dir.glob('tasks/*.rake').each { |r| import r }
@@ -30,7 +29,7 @@ task spec: :rubocop do
 end
 
 # Could not find jetty:config within release 2.0.3
-task ci: ['engine_cart:generate', 'jetty:clean', 'jetty:config', 'db:test:prepare'] do
+task ci: ['jetty:clean', 'jetty:config', 'db:test:prepare'] do
   puts 'running continuous integration'
   jetty_params = Jettywrapper.load_config
   jetty_params[:startup_wait] = 90
@@ -41,5 +40,4 @@ task ci: ['engine_cart:generate', 'jetty:clean', 'jetty:config', 'db:test:prepar
   fail "test failures: #{error}" if error
 end
 
-task clean: 'engine_cart:clean'
 task default: :ci
