+++ geo_concerns/app/models/concerns/feature_extraction_behavior.rb
@@ -1,0 +1,8 @@
+
+module FeatureExtractionBehavior
+  extend ActiveSupport::Concern
+  include ::CurationConcerns::GenericWorkBehavior
+  include ::CurationConcerns::BasicMetadata
+  include ::BasicGeoMetadata
+  include ::GeoreferencedBehavior
+end
