--- geo_concerns/app/models/concerns/geo_concerns/feature_extraction_behavior.rb
@@ -1,14 +1,0 @@
-module GeoConcerns
-  module FeatureExtractionBehavior
-    extend ActiveSupport::Concern
-    include ::CurationConcerns::GenericWorkBehavior
-    include ::CurationConcerns::BasicMetadata
-    include ::GeoConcerns::BasicGeoMetadata
-    include ::GeoConcerns::GeoreferencedBehavior
-
-    # associated_with :raster (derived from, required)
-
-    # property :transcribedBy, String
-    # ...
-  end
-end
