--- geo_concerns/app/models/concerns/geo_concerns/image_behavior.rb
@@ -1,31 +1,0 @@
-module GeoConcerns
-  module ImageBehavior
-    extend ActiveSupport::Concern
-    include ::CurationConcerns::GenericWorkBehavior
-    include ::CurationConcerns::BasicMetadata
-    include ::GeoConcerns::BasicGeoMetadata
-
-    included do
-      aggregates :rasters, predicate: RDF::Vocab::ORE.aggregates,
-                           class_name: 'GeoConcerns::Vector',
-                           type_validator: type_validator
-      filters_association :members, as: :image_files, condition: :concerns_image_file?
-    end
-
-    def concerns_image?
-      true
-    end
-
-    def concerns_image_file?
-      false
-    end
-
-    def image_file
-      image_files.first
-    end
-
-    def image_file=(_file)
-      image_files=([_file])
-    end
-  end
-end
