--- geo_concerns/app/models/concerns/geo_concerns/raster_file_behavior.rb
@@ -1,23 +1,0 @@
-module GeoConcerns
-  module RasterFileBehavior
-    extend ActiveSupport::Concern
-    include Hydra::Works::GenericFileBehavior # Added
-    include ::CurationConcerns::GenericFileBehavior
-    include ::GeoConcerns::BasicGeoMetadata
-    include ::GeoConcerns::GeoreferencedBehavior
-
-    # @return [Boolean] whether this instance is a GeoConcerns Raster File.
-    def concerns_raster_file?
-      true
-    end
-
-    # Aliases the thumbnail method
-    def preview
-      thumbnail
-    end
-
-    def raster
-      parents.find { |parent| parent.class.included_modules.include?(GeoConcerns::RasterBehavior) }
-    end
-  end
-end
