--- geo_concerns/app/models/concerns/geo_concerns/vector_metadata_file_behavior.rb
@@ -1,12 +1,0 @@
-module GeoConcerns
-  module VectorMetadataFileBehavior
-    extend ActiveSupport::Concern
-    include Hydra::Works::GenericFileBehavior
-    include ::CurationConcerns::GenericFileBehavior
-    include ::GeoConcerns::MetadataFileBehavior
-
-    def vector
-      aggregated_by.find { |parent| parent.class.included_modules.include?(GeoConcerns::VectorBehavior) }
-    end
-  end
-end
