+++ geo_concerns/app/models/concerns/vector_metadata_file_behavior.rb
@@ -1,0 +1,11 @@
+
+module VectorMetadataFileBehavior
+  extend ActiveSupport::Concern
+  include Hydra::Works::GenericFileBehavior
+  include ::CurationConcerns::GenericFileBehavior
+  include ::MetadataFileBehavior
+
+  def vector
+    aggregated_by.find { |parent| parent.class.included_modules.include?(GeoConcerns::VectorBehavior) }
+  end
+end
