+++ geo_concerns/app/models/image_file.rb
@@ -1,4 +1,4 @@
-# Generated by curation_concerns:models:install
+
 class ImageFile < ActiveFedora::Base
   include ::GeoConcerns::ImageFileBehavior
 end
