+++ geo_concerns/geo_concerns.gemspec
@@ -1,15 +1,18 @@
 # Coding: utf-8
+version = File.read(File.expand_path("../VERSION", __FILE__)).strip
 
-# This is only in place in order to pass the project name to engine_cart
 Gem::Specification.new do |spec|
   spec.name          = "geo_concerns"
-
+  spec.version       = version
   spec.authors       = ["Darren Hardy", "Eliot Jordan", "John Huck", "Eric James", "James Griffin"]
-
-  spec.summary       = %q{Hydra GIS data models in compliance with the PCDM }
-  spec.description   = %q{Hydra GIS data models in compliance with the PCDM }
+  spec.email         = [""]
+  spec.summary       = %q{Rails application for developing Hydra Geo models. Built around Curation Concerns engine. }
+  spec.description   = %q{Rails application for developing Hydra Geo models. Built around Curation Concerns engine. }
   spec.homepage      = ""
   spec.license       = "APACHE2"
 
+  spec.files         = `git ls-files -z`.split("\x0")
+  spec.executables   = spec.files.grep(%r{^bin/}) { |f| File.basename(f) }
   spec.test_files    = spec.files.grep(%r{^(test|spec|features)/})
+  spec.require_paths = []
 end
