+++ geo_concerns/spec/models/raster_file_spec.rb
@@ -133,8 +133,10 @@ describe RasterFile do
                                  crs: 'urn:ogc:def:crs:EPSG::6326').to_solr
     }
 
-    it "indexes bbox field" do
+    it "indexes the bounding box" do
       expect(solr_doc.keys).to include 'georss_box_tesim'
+    end
+    it "indexes the coordinate reference system" do
       expect(solr_doc.keys).to include 'crs_tesim'
     end
   end
